ai_service/requirements.txt
fastapi
uvicorn
langchain
chromadb
pydantic
python-dotenv
langchain-community
ollama